+++ test/Rector/Class_/ImplementsFactoryInterfaceToPsrFactoryRector/ImplementsFactoryInterfaceToPsrFactoryRectorTest.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace LaminasTest\ServiceManager\Migration\Rector\Class_\ImplementsFactoryInterfaceToToPsrFactoryRector;
+namespace LaminasTest\ServiceManager\Migration\Rector\Class_\ImplementsFactoryInterfaceToPsrFactoryRector;
 
 use Iterator;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 use Symplify\SmartFileSystem\SmartFileInfo;
 
-final class ImplementsFactoryInterfaceToToPsrFactoryRectorTest extends AbstractRectorTestCase
+final class ImplementsFactoryInterfaceToPsrFactoryRectorTest extends AbstractRectorTestCase
 {
     /**
      * @dataProvider provideData()
