+++ test/Rector/Class_/ImplementsFactoryInterfaceToPsrFactoryRector/config/configured_rule.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-use Laminas\ServiceManager\Migration\Rector\Class_\ImplementsFactoryInterfaceToToPsrFactoryRector;
+use Laminas\ServiceManager\Migration\Rector\Class_\ImplementsFactoryInterfaceToPsrFactoryRector;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
     $services = $containerConfigurator->services();
-    $services->set(ImplementsFactoryInterfaceToToPsrFactoryRector::class);
+    $services->set(ImplementsFactoryInterfaceToPsrFactoryRector::class);
 };
